--- components/ai_chat/resources/shared_conversation_lib/create_nala_style_element.ts
@@ -1,12 +1,0 @@
-// Copyright (c) 2026 The Brave Authors. All rights reserved.
-// This Source Code Form is subject to the terms of the Mozilla Public
-// License, v. 2.0. If a copy of the MPL was not distributed with this file,
-// You can obtain one at https://mozilla.org/MPL/2.0/.
-
-export async function createNalaStyleElement(shadow: ShadowRoot) {
-  const style = document.createElement('style')
-  const scriptUrl = import.meta.url
-  const relativePathUrl = new URL('./nala.css', scriptUrl)
-  style.textContent = await fetch(relativePathUrl).then((res) => res.text())
-  shadow.appendChild(style)
-}
