+++ components/ai_chat/resources/shared_conversation_lib/webpack.config.ts
@@ -110,7 +110,7 @@ export default async function (
       ),
       to: 'nala.css',
       transform: {
-        // TODO(https://github.com/brave/leo/issues/1433) Remove this transform
+        // TODO(https://github.com/brave/leo/issues/1433): Remove this transform
         // when nala supports `:host`.
         transformer(content: Buffer, absoluteFrom: string) {
           return content.toString().replaceAll(':root', ':host')
